+++ scripts/database/otrs-schema.mysql.sql
@@ -1226,20 +1226,6 @@ CREATE TABLE gi_debugger_entry_content (
     INDEX gi_debugger_entry_content_debug_level (debug_level)
 );
 # ----------------------------------------------------------
-#  create table gi_object_lock_state
-# ----------------------------------------------------------
-CREATE TABLE gi_object_lock_state (
-    webservice_id INTEGER NOT NULL,
-    object_type VARCHAR (30) NOT NULL,
-    object_id BIGINT NOT NULL,
-    lock_state VARCHAR (30) NOT NULL,
-    lock_state_counter INTEGER NOT NULL,
-    create_time DATETIME NOT NULL,
-    change_time DATETIME NOT NULL,
-    UNIQUE INDEX gi_object_lock_state_list (webservice_id, object_type, object_id),
-    INDEX object_lock_state_list_state (webservice_id, object_type, object_id, lock_state)
-);
-# ----------------------------------------------------------
 #  create table smime_signer_cert_relations
 # ----------------------------------------------------------
 CREATE TABLE smime_signer_cert_relations (